Slightly corrects style errors in the Cartridge hierarchy, and introduces mapping of .ROM to the MSX when appropriate.

+++ Storage/Cartridge/Cartridge.cpp
@@ -8,4 +8,5 @@
 
 #include "Cartridge.hpp"
 
-const int Storage::Cartridge::Cartridge::Segment::UnknownAddress = -1;
+const size_t Storage::Cartridge::Cartridge::Segment::UnknownAddress = -1;
+
